requirements.txt
appdirs
CacheControl
certifi
chardet
colorama
contextlib2
dill
distlib
distro
dominate
html5lib
idna
ipaddr
lockfile
msgpack
numpy
packaging
pep517
Pillow
progress
pyparsing
pytoml
requests
retrying
six
torch
torchvision
typing-extensions
urllib3
webencodings